requirements.txt
newspaper3k
ScrapyElasticSearch

